Added recyclerview.onScrollListener.
This way the scroll state of the inner recyclerviews are saved when the user scrolls and not only when they tap on a certain bus. Some corrections on comments and a clearer onBindViewHolder code (I think). Also erased lateral padding on the layout on the stops so the recyclerview get to the borders of the screen while scrolling, I guess it looks better this way.