New files with the necessary data.
Added a new model and a new class to hold a map with all the stops and their info, even though that implementation might not be final, just testing it. Also we can now use that data to populate the UI and check everything, but I only did it with a few stops so let's see. TL/DR: everything new in this commit is in a test stage.