Changed how the detail layout expands.
Now, instead of relying on the notifyItemChanged() animation we have a expand animation that looks much more clean and we don't need to recreate that view. This broke the detail change on tap but fix is on the next commit. Also, little changes to the layout and every text set to black. Some fixes in names in AppData, with more info on the lines.