LiveData implementation.
More work on the LiveData implementation, now it works. We even have a loading view on the buses layout! I just realized that the ModelViews from the last commits aren't really neccesary so we will just forget about them on the next commits, along with some clean-up so I don't go crazy reading this.