Little layout change.
Now, the distance to the stop is below the name to avoid overlapping, along with how much time has been since the data was updated. Also, I noticed that you can't directly change the font of a TickerView, so we have to make it dinamically (a bit strange but at least we can).