More Layout changes.
Now the units on the updated time and the distance info are also TickerView. Fixed the dot size and the full sentence thing. Also, very component is in a new layout there to make it easy to read and develop further. More changes will be needed but right now we have a pretty close look to the final version.