More refactor!
This is a serious one. I learned that we can update the recyclerview elements without a full redraw (maybe dumb to the reader), so I realized that is waaaay better to do that instead of the mess with liveData we had before. Now, the observing takes place on the fragment, which passes the updates to the DiffUtils methods on the adapter. This way, UI and logic stay away from each other. Also some superfluous code has been deleted and a new Factory to initialize the ViewModel.