More layout work!
First, now the detail data is shown and updated. Added the units for the distance left, with a little change on BusModelView to avoid unnecessary checks on the UI, we will do the cleaning work on the repository. Also, time left units are now a TickerView. And last, I set a longer time for all the ticker animations, as I feel they were to fast to be pretty; not definitive, but way nicer without being boring.