Several things.
First, now the fragment ViewModel will be linked to the activity lifecycle; if not, it is recreated with the fragment on rotations, will defeats the purpose of it. In a future commit there will be an updated state ViewModel with the state of both recyclerviews.

And there's now ground work for fetching data from the web we now have permissions and a retrofit instance, even though it just outputs to the log. I also made some utils to parse the raw data from the web, that needs a bit of cleaning but nothing fully implemented yet.