Units refactor.
Now, instead of having two varibales in every model we have a pair of magnitude and unit. This way, we can evaluate in the UI the language of the unit, and overall I think it's a more elegant code, as a magnitude always has to have a unit.

I dind't find a way to put the Pair on a Bundle, so their components are divided at the DiffUtil evaluation. I will keep finding a way, I don't like it like this even though it works.